+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # [develop](https://github.com/adhearsion/adhearsion-asr)
 
+# [v1.0.1](https://github.com/adhearsion/adhearsion-asr/compare/1.0.0...1.0.1) - [2013-09-30](https://rubygems.org/gems/adhearsion-asr/versions/1.0.1)
+  * Bugfix: A menu definition's block context is now available
+
 # [v1.0.0](https://github.com/adhearsion/adhearsion-asr/compare/0.1.0...1.0.0) - [2013-08-29](https://rubygems.org/gems/adhearsion-asr/versions/1.0.0)
   * Change: Controller methods are now included in all call controllers by default, but this is configurable
   * Change: Default renderer/voice config is moved to Adhearsion core